fix: add pages to the template
